refactor(dispatch): extract capability-resolution heuristics from router.py; fix stale docstring (DR-12)
Moves resolve_by_capability into dispatcher/_capability_resolve.py with a
thin import back; corrects the router.py gate docstring that described a
private _current_state reach-in the code no longer does. Behavior-preserving.
